(WIP) rework the sessions calendar into a more modular form
