+++ src/components/SessionsCalendar/DayComponent/types.ts
@@ -1,5 +1,7 @@
 import type {DayState, DateData} from 'react-native-calendars/src/types';
 
+type CalendarColors = 'yellow' | 'red' | 'orange' | 'black' | 'green';
+
 type DayMarking = {
   units?: number;
   color?: string;
@@ -14,4 +16,4 @@ type DayComponentProps = {
   onPress?: (day: DateData) => void;
 };
 
-export type {DayMarking, DayComponentProps};
+export type {DayMarking, DayComponentProps, CalendarColors};
