+++ src/screens/HomeScreen.tsx
@@ -7,20 +7,18 @@ import {
   View,
 } from 'react-native';
 import MenuIcon from '@components/Buttons/MenuIcon';
-import SessionsCalendar from '@components/SessionsCalendar/SessionsCalendar';
+import SessionsCalendar from '@components/SessionsCalendar';
 import type {DateData} from 'react-native-calendars';
 import * as KirokuIcons from '@components/Icon/KirokuIcons';
 import {
-  dateToDateObject,
   calculateThisMonthDrinks,
   calculateThisMonthUnits,
   timestampToDate,
-  timestampToDateString,
+  dateToDateData,
 } from '@libs/DataHandling';
 import {useUserConnection} from '@context/global/UserConnectionContext';
 import UserOffline from '@components/UserOfflineModal';
 import {synchronizeUserStatus} from '@database/users';
-import commonStyles from '@src/styles/commonStyles';
 import {useFirebase} from '@context/global/FirebaseContext';
 import ProfileImage from '@components/ProfileImage';
 import CONST from '@src/CONST';
@@ -52,12 +50,9 @@ import useLocalize from '@hooks/useLocalize';
 import {roundToTwoDecimalPlaces} from '@libs/NumberUtils';
 import NoSessionsInfo from '@components/NoSessionsInfo';
 import Text from '@components/Text';
-import variables from '@src/styles/variables';
-import stylePropTypes from '@src/styles/stylePropTypes';
 import useWindowDimensions from '@hooks/useWindowDimensions';
 
 type State = {
-  visibleDateObject: DateData;
   drinkingSessionsCount: number;
   drinksConsumed: number;
   unitsConsumed: number;
@@ -70,7 +65,6 @@ type Action = {
 };
 
 const initialState: State = {
-  visibleDateObject: dateToDateObject(new Date()),
   drinkingSessionsCount: 0,
   drinksConsumed: 0,
   unitsConsumed: 0,
@@ -79,8 +73,6 @@ const initialState: State = {
 
 const reducer = (state: State, action: Action): State => {
   switch (action.type) {
-    case 'SET_VISIBLE_DATE_OBJECT':
-      return {...state, visibleDateObject: action.payload};
     case 'SET_DRINKING_SESSIONS_COUNT':
       return {...state, drinkingSessionsCount: action.payload};
     case 'SET_DRINKS_CONSUMED':
@@ -114,6 +106,9 @@ function HomeScreen({route}: HomeScreenProps) {
     userData,
     isLoading,
   } = useDatabaseData();
+  const [visibleDate, setVisibleDate] = useState<DateData>(
+    dateToDateData(new Date()),
+  );
   const [ongoingSessionId, setOngoingSessionId] = useState<
     DrinkingSessionId | null | undefined
   >();
@@ -172,16 +167,16 @@ function HomeScreen({route}: HomeScreenProps) {
       ? Object.values(drinkingSessionData)
       : [];
     const thisMonthDrinks = calculateThisMonthDrinks(
-      state.visibleDateObject,
+      visibleDate,
       drinkingSessionArray,
     );
     const thisMonthUnits = calculateThisMonthUnits(
-      state.visibleDateObject,
+      visibleDate,
       drinkingSessionArray,
       preferences.drinks_to_units,
     );
     const thisMonthSessionCount = DSUtils.getSingleMonthDrinkingSessions(
-      timestampToDate(state.visibleDateObject.timestamp),
+      timestampToDate(visibleDate.timestamp),
       drinkingSessionArray,
       false,
     ).length; // Replace this in the future
@@ -192,7 +187,7 @@ function HomeScreen({route}: HomeScreenProps) {
     });
     dispatch({type: 'SET_DRINKS_CONSUMED', payload: thisMonthDrinks});
     dispatch({type: 'SET_UNITS_CONSUMED', payload: thisMonthUnits});
-  }, [drinkingSessionData, state.visibleDateObject, preferences]);
+  }, [drinkingSessionData, visibleDate, preferences]);
 
   useMemo(() => {
     const sessionsAreMissingTz =
@@ -298,14 +293,10 @@ function HomeScreen({route}: HomeScreenProps) {
             </View>
             <SessionsCalendar
               userID={user.uid}
+              visibleDate={visibleDate}
+              onDateChange={(date: DateData) => setVisibleDate(date)}
               drinkingSessionData={drinkingSessionData}
               preferences={preferences}
-              onDateChange={(date: Date) => {
-                dispatch({
-                  type: 'SET_VISIBLE_DATE_OBJECT',
-                  payload: dateToDateObject(date),
-                });
-              }}
             />
           </>
         ) : (
